Send the plainTextValue and mentions array arguments to the handleChange callback when adding a new mention
Currently these arguments are `undefined` in my `handleChange` action, where on `keyDown` they are not. This means when a user's last input was inserting a mention, it is not realized until they do an additional `keyDown` action. This change mimics the behavior done on `keyDown`, and sends the additional function parameters.